Add Enemy AI, Healing player, update damage system

+++ CubeEvolution2/Assets/Scripts/Gameplay/Chatacter/Character.cs
@@ -20,17 +20,17 @@ public class Character : ScriptableObject
     public int UnlockPrice { get => _unlockPrice; }
 
     [Header("Character Skills")]
-    [SerializeField] private float _health;
-    [SerializeField] private float _damage;
+    [SerializeField] private int _health;
+    [SerializeField] private int _damage;
     [SerializeField] private float _speed;
-    [SerializeField] private float _healthMax;
-    [SerializeField] private float _damageMax;
+    [SerializeField] private int _healthMax;
+    [SerializeField] private int _damageMax;
     [SerializeField] private float _speedMax;
-    public float Health { get => _health; set => _health = value;}
-    public float Damage { get => _damage; set => _damage = value;}
+    public int Health { get => _health; set => _health = value;}
+    public int Damage { get => _damage; set => _damage = value;}
     public float Speed { get => _speed; set => _speed = value;}
-    public float HealthMax { get => _healthMax; }
-    public float DamageMax { get => _damageMax; }
+    public int HealthMax { get => _healthMax; }
+    public int DamageMax { get => _damageMax; }
     public float SpeedMax { get => _speedMax; }
     
     public static float StatsMax { get; } = 2000f;
